supports multiple toolbox instances
