+++ lib/options.js
@@ -24,17 +24,19 @@ if (self.registration) {
   scope = self.scope || new URL('./', self.location).href;
 }
 
-module.exports = {
-  cache: {
-    name: '$$$toolbox-cache$$$' + scope + '$$$',
-    maxAgeSeconds: null,
-    maxEntries: null
-  },
-  debug: false,
-  networkTimeoutSeconds: null,
-  preCacheItems: [],
-  // A regular expression to apply to HTTP response codes. Codes that match
-  // will be considered successes, while others will not, and will not be
-  // cached.
-  successResponses: /^0|([123]\d\d)|(40[14567])|410$/
+module.exports = function () {
+  return {
+    cache: {
+      name: '$$$toolbox-cache$$$' + scope + '$$$',
+      maxAgeSeconds: null,
+      maxEntries: null
+    },
+    debug: false,
+    networkTimeoutSeconds: null,
+    preCacheItems: [],
+    // A regular expression to apply to HTTP response codes. Codes that match
+    // will be considered successes, while others will not, and will not be
+    // cached.
+    successResponses: /^0|([123]\d\d)|(40[14567])|410$/
+  }
 };
