+++ lib/router.js
@@ -130,4 +130,6 @@ Router.prototype.match = function(request) {
       this.matchMethod('any', request.url);
 };
 
-module.exports = new Router();
+module.exports = function () {
+  return new Router();
+}
